@@ -1,7 +1,7 @@
 {
   "name": "mmf-blog-api-ts",
   "version": "1.0.1",
-  "packageManager": "pnpm@9.12.3",
+  "packageManager": "pnpm@9.14.2",
   "description": "mmf blog api server",
   "author": "lincenying &lt;lincenying@qq.com&gt;",
   "license": "MIT",
@@ -18,8 +18,8 @@
     "deploy-watch": "nodemon --delay 500ms --watch ./dist dist/app.js",
     "serve": "nodemon --watch ./src --exec tsx ./src/app.ts",
     "start": "npm run deploy",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
+    "lint": "eslint . --flag unstable_ts_config",
+    "lint:fix": "eslint . --flag unstable_ts_config --fix",
     "lint:ts": "tsc --noEmit"
   },
   "dependencies": {
@@ -40,18 +40,18 @@
     "markdown-it-toc-and-anchor": "^4.2.0",
     "md5": "^2.3.0",
     "mockjs": "^1.1.0",
-    "mongoose": "^8.0.4",
+    "mongoose": "^8.8.2",
     "mongoose-autopopulate": "^1.1.0",
     "morgan": "^1.10.0",
     "multer": "1.4.5-lts.1",
-    "qiniu": "^7.13.0",
+    "qiniu": "^7.14.0",
     "request-ip": "^3.3.0",
     "serve-favicon": "^2.5.0",
     "tslib": "^2.8.1",
     "twig": "^1.17.1"
   },
   "devDependencies": {
-    "@lincy/eslint-config": "^5.2.7",
+    "@lincy/eslint-config": "^5.3.0",
     "@types/compression": "^1.7.5",
     "@types/cookie-parser": "^1.4.7",
     "@types/express": "^5.0.0",
@@ -61,11 +61,12 @@
     "@types/mockjs": "^1.0.10",
     "@types/morgan": "^1.9.9",
     "@types/multer": "^1.4.12",
-    "@types/node": "^22.9.0",
+    "@types/node": "^22.9.3",
     "@types/request-ip": "^0.0.41",
     "@types/serve-favicon": "^2.5.7",
     "concurrently": "^9.1.0",
-    "eslint": "^9.14.0",
+    "eslint": "^9.15.0",
+    "jiti": "^2.4.0",
     "lint-staged": "^15.2.10",
     "nodemon": "^3.1.7",
     "simple-git-hooks": "^2.11.1",
@@ -85,7 +86,7 @@
   },
   "lint-staged": {
     "*.{ts,js}": [
-      "eslint --fix"
+      "eslint --flag unstable_ts_config --fix"
     ]
   }
 }
